refactor: shift `RunConfig` back to `debiasing/debias_wrapper.py`
