+++ python/clim_recal/utils/core.py
@@ -43,6 +43,10 @@ CURRENT_PATH = Path().absolute()
 PYTHON_PACKAGE_DIR_NAME: Final[Path] = Path("python")
 RUN_TIME_STAMP_FORMAT: Final[str] = "%y-%m-%d_%H-%M"
 
+DEBIAN_MOUNT_PATH: Final[Path] = Path("/mnt/vmfileshare")
+DARWIN_MOUNT_PATH: Final[Path] = Path("/Volumes/vmfileshare")
+CLIMATE_DATA_PATH: Final[Path] = Path("ClimateData")
+
 
 @dataclass
 class MonthDay:
@@ -701,6 +705,48 @@ def csv_reader(path: PathLike, **kwargs) -> Iterator[dict[str, str]]:
     with open(path) as csv_file:
         for row in DictReader(csv_file, **kwargs):
             yield row
+
+
+def climate_data_mount_path(
+    is_darwin: bool | None = None, full_path: bool = True
+) -> Path:
+    """Return likely climate data mount path based on operating system.
+
+    Parameters
+    ----------
+    is_darwin
+        Whether to use `CLIMATE_DATA_MOUNT_PATH_DARWIN` or
+        call `is_platform_darwin` if None. fixture `is_platform_darwin`.
+
+    Returns
+    -------
+    The `Path` climate data would likely be mounted to.
+    """
+    path: Path
+    if is_darwin is None:
+        is_darwin = is_platform_darwin()
+    if is_darwin:
+        path = DARWIN_MOUNT_PATH
+    else:
+        path = DEBIAN_MOUNT_PATH
+    if full_path:
+        return path / CLIMATE_DATA_PATH
+    else:
+        return path
+
+
+def is_climate_data_mounted(mount_path: PathLike | None = None) -> bool:
+    """Check if `CLIMATE_DATA_MOUNT_PATH` is mounted.
+
+    Notes
+    -----
+    Consider refactoring to `climate_data_mount_path` returns `None`
+    when conditions here return `False`.
+    """
+    if not mount_path:
+        mount_path = climate_data_mount_path()
+    assert isinstance(mount_path, Path)
+    return mount_path.exists()
 
 
 class StrEnumReprName(StrEnum):
